requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==4.2.7
django-cors-headers==4.3.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.45.1
frozenlist==1.4.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
multidict==6.0.4
packaging==23.2
phonenumberslite==8.13.27
Pillow==10.1.0
pycparser==2.21
pydyf==0.8.0
PyJWT==2.8.0
pyphen==0.14.0
pytz==2023.3.post1
reportlab==4.0.7
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
twilio==8.10.1
typing_extensions==4.9.0
urllib3==2.1.0
weasyprint==60.1
webencodings==0.5.1
yarl==1.9.2
zopfli==0.2.3
